Hii,
I am excited to share with you the Tableau file for our NBA Forecast project. This file includes all of the data visualization and analysis I have done to forecast the performance of the teams by using Tableau. Please take a look and let meknow if you have any questions or suggestions.